requirements.txt
attrs==24.2.0
black==23.7.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
flake8==6.1.0
h11==0.14.0
idna==3.7
mccabe==0.7.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pycodestyle==2.11.0
pyflakes==3.1.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
selenium==4.23.1
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.26.1
trio-websocket==0.11.1
typing-extensions==4.12.2
urllib3==2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
